Card Hash movido para o CardService
