+++ Pagarme/src/main/java/br/com/aspotato/pagarme/services/CardService.java
@@ -5,7 +5,10 @@ import com.mashape.unirest.http.JsonNode;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.request.HttpRequestWithBody;
 import br.com.aspotato.pagarme.models.Card;
+import br.com.aspotato.pagarme.models.Key;
 import br.com.aspotato.pagarme.utils.PagarMeProvider;
+import br.com.aspotato.pagarme.utils.PagarMeUtil;
+import org.json.JSONObject;
 
 public class CardService extends BasicService {
 
@@ -25,5 +28,18 @@ public class CardService extends BasicService {
 
 		return jsonResponse;
 	}
+        
+        
+        public Key generateCardHash() throws Exception {
+            HttpResponse<JsonNode> jsonResponse = Unirest.get(instance.getUrl() + "1/transactions/card_hash_key")
+                            .header("accept", "application/json")
+                            .queryString("encryption_key", instance.getEncryptionKey())
+                            .asJson();
+            JSONObject resultObject = jsonResponse.getBody().getObject();
+            
+            this.checkErrors(resultObject);
+            return (Key) PagarMeUtil.convertJsonToObject(Key.class, resultObject);
+            
+        }
 
 }
